// packages/prisma-oppulence/generated/zod/inputTypeSchemas/PipelineAutomationMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const PipelineAutomationMaxOrderByAggregateInputSchema: z.ZodType<Prisma.PipelineAutomationMaxOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        pipelineId: z.lazy(() => SortOrderSchema).optional(),
        name: z.lazy(() => SortOrderSchema).optional(),
        description: z.lazy(() => SortOrderSchema).optional(),
        isActive: z.lazy(() => SortOrderSchema).optional(),
        triggerType: z.lazy(() => SortOrderSchema).optional(),
        actionType: z.lazy(() => SortOrderSchema).optional(),
        priority: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        updatedAt: z.lazy(() => SortOrderSchema).optional(),
        lastTriggered: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default PipelineAutomationMaxOrderByAggregateInputSchema